requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1

# PDF parsing
pymupdf==1.24.14

# Data
pandas==2.2.3
pyarrow==18.1.0

# Embeddings + Vector search
sentence-transformers==3.2.1
faiss-cpu==1.9.0.post1

# Utilities
tqdm==4.67.1
numpy<2.0
